added snapped and balkanized keywords, changed scripts to check for them as needed. --Molly
git-svn-id: https://giordano.mit.edu/nvs/mangle/trunk/dev_80bit@91 9f78a35e-8a50-488b-8db1-b351ce30c53d

+++ masks/healpix/healpixrast.sh
@@ -67,17 +67,17 @@ res1=`awk '{print substr($2, 1, length($2)-1)}' < jpix`
 scheme1=`awk '{print substr($2, length($2))}' < jpix`
 rm jpix
 
-#if input file is unpixelized, pixelize it and snap it
+#check if input file is snapped and balkanized
+snapped=`awk '/snapped/{print $1}' < $1`
+balkanized=`awk '/balkanized/{print $1}' < $1`
+
+#if input file is unpixelized, pixelize it
 #if input file is pixelized to the correct resolution, use it as is.
-#WARNING: this assumes that if the input polygon file is pixelized, it is also snapped.
 if [ "$res1" = "" ]; then
     echo ""
     echo "Pixelizing $1 ..."
     $MANGLEBINDIR/pixelize -P${scheme}0,$pix $1 jp || exit   
     echo ""
-    echo "Snapping $1 ..."
-    $MANGLEBINDIR/snap jp jps || exit
-    rm jp
 #if input file is pixelized to a different resolution and scheme,
 #warn and pixelize again
 elif [ ! "$res1" = "$pix" ] || [ ! "$scheme1" = "$scheme" ]; then
@@ -90,12 +90,30 @@ elif [ ! "$res1" = "$pix" ] || [ ! "$scheme1" = "$scheme" ]; then
     echo "Pixelizing $1 ..."
     $MANGLEBINDIR/pixelize -P${scheme}0,$pix $1 jps || exit   
 else
-    cp $1 jps
+    cp $1 jp
+fi
+
+#if input file isn't snapped, snap it
+if [ ! "$snapped" = "snapped" ]; then
+    echo "Snapping $1 ..."
+    $MANGLEBINDIR/snap jp jps || exit
+    rm jp
+else
+    mv jp jps
+fi
+
+#if input file isn't balkanized, balkanize it
+if [ ! "$balkanized" = "balkanized" ]; then
+    echo "Balkanizing $1 ..."
+    $MANGLEBINDIR/balkanize jps jb || exit
+    rm jps
+else
+    mv jps jb
 fi
 
 echo ""
 echo "Rasterizing $1 against the Nside=$2 approximate HEALPix pixels..."
-$MANGLEBINDIR/rasterize $healpixfile jps $3 || exit
-rm jps
+$MANGLEBINDIR/rasterize $healpixfile jb $3 || exit
+rm jb
 
 echo "Rasterized mask written to $3."
